Moved moduleLogger to module/internal
